oura_demo/api/pyproject.toml
[project]
name = "oura-demo-api"
version = "0.1.0"
description = "Lightweight demo API for Oura time series synthesis"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "pandas>=2.0.0",
    "pyarrow>=14.0.0",
    "numpy>=1.24.0",
    "pydantic>=2.0.0",
    "python-multipart>=0.0.9",
    "openai>=1.0.0",
    "pyyaml>=6.0.0",
    "loguru>=0.7.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0", "httpx>=0.27.0"]
# synthefy_pkg is installed separately (either via pip install -e or in Docker)

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
